requirements.txt
fastapi[standard]>=0.68.0
uvicorn>=0.15.0
motor>=3.0.0
pydantic>=1.8.0
python-dotenv>=0.19.0
boto3>=1.26.0
botocore>=1.29.0
pymongo>=4.0.0
python-multipart>=0.0.5
beanie~=1.29.0
pylint>=2.11.1
black>=21.12b0
starlette~=0.46.0
pandas>=1.3.0
openpyxl>=3.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4